renamed module bookkeeping_logic to util.bookings
